requirements.txt
Django==2.1.5
djangorestframework==3.8.2
django-cors-headers==2.4.0
django-environ==0.4.5
psycopg2-binary==2.8.5
dj-database-url==0.5.0
gunicorn==19.9.0
whitenoise==4.1.2
django-environ==0.4.5
